Improve narrow phase collision detection
narrow pase collision detection now uses a simple test
based on the dimensions of the sprite to determine if object collide.

Won't be accurate if sprites rotate. or if the the visible portion
of the sprite is significantly smaller than the edge of the texture.

Won't look accurrate for sprites that don't look squarish